+++ src/main/java/ru/netology/stats/SalesStatistics.java
@@ -2,48 +2,48 @@ package ru.netology.stats;
 
 public class SalesStatistics {
 
-    public long sumOfSales(long[] sales) {
+    public int sumOfSales(int[] sales) {
         long sum = 0; // Общие продажи
-        for (long i = 0; i < sales.length; i++) {
-            sum += sales[(int) i];
+        for (int i = 0; i < sales.length; i++) {
+            sum += sales[i];
         }
-        return sum;
+        return (int) sum;
     }
 
-    public long averageSumOfSales(long[] sales) {
+    public int averageSumOfSales(int[] sales) {
         long sum = 0;
         long averageSum = 0;  // Средние продажи
         for (long i = 0; i < sales.length; i++) {
             sum += sales[(int) i];
             averageSum = (sum / sales.length);
         }
-        return averageSum;
+        return (int) averageSum;
     }
 
 
-    public long numMonthOfMaxSales(long[] sales) {
-        int maxMonth = 0; // Месяцев Максимальных продаж
+    public int numMonthOfMaxSales(int[] sales) {
+        long maxMonth = 0; // Месяцев Максимальных продаж
         for (long i = 0; i < sales.length; i++) {
             if (sales[(int) i] >= sales[(int) maxMonth]) {
                 maxMonth = (int) i;
             }
         }
-        return maxMonth + 1;
+        return (int) (maxMonth + 1);
     }
 
-    public long numMonthOfMinSales(long[] sales) {
-        int minMonth = 0; // Месяцев Минимальных продаж
+    public int numMonthOfMinSales(int[] sales) {
+        long minMonth = 0; // Месяцев Минимальных продаж
         for (long i = 0; i < sales.length; i++) {
             if (sales[(int) i] <= sales[(int) minMonth]) {
                 minMonth = (int) i;
             }
         }
-        return minMonth + 1;
+        return (int) (minMonth + 1);
     }
 
-    public long countMonthOfSalesBelowAverage(long[] sales) {
+    public int countMonthOfSalesBelowAverage(int[] sales) {
         int count = 0;
-        long averageSum = averageSumOfSales(sales); // Месяцев продаж ниже среднего
+        int averageSum = averageSumOfSales(sales); // Месяцев продаж ниже среднего
         for (long i = 0; i < sales.length; i++) {
             if (sales[(int) i] > averageSum) {
                 count++;
@@ -52,9 +52,9 @@ public class SalesStatistics {
         return count;
     }
 
-    public long countMonthOfSalesAboveAverage(long[] sales) {
+    public int countMonthOfSalesAboveAverage(int[] sales) {
         int count = 0;
-        long averageSum = averageSumOfSales(sales); // Месяцев продаж выше среднего
+        int averageSum = averageSumOfSales(sales); // Месяцев продаж выше среднего
         for (long i = 0; i < sales.length; i++) {
             if (sales[(int) i] < averageSum) {
                 count++;
